InterSpikeSpectra

Documentation for InterSpikeSpectra.

InterSpikeSpectra.compute_spectrum_according_to_actual_spectrum

InterSpikeSpectra.compute_spectrum_according_to_threshold3

InterSpikeSpectra.create_single_basis_function

InterSpikeSpectra.generate_basis_functions

InterSpikeSpectra.inter_spike_spectrum

InterSpikeSpectra.inter_spike_spectrum2

InterSpikeSpectra.inter_spike_spectrum3

InterSpikeSpectra.compute_spectrum_according_to_actual_spectrum — Method

Determine the right regularization parameter with respect to ρ_thres & tol
using the Newton-method.

InterSpikeSpectra.compute_spectrum_according_to_threshold3 — Method

Determine the right regularization parameter with respect to ρ_thres & tol
using the Newton-method.

InterSpikeSpectra.create_single_basis_function — Method

Create a spike-basis functions of length `N` and inter-spike-interval
`period`. Due to time shifts, there are `period` different, linearly dependent
basis functions.

InterSpikeSpectra.generate_basis_functions — Method

generate_basis_functions(N::Int) → basis

Generate a matrix `basis` with all possible `sum(1:N)` spike-basis functions.
These functions are not independent, since we account for time shifts. Thus,
`basis` is a set of linear-dependent basis- functions.

InterSpikeSpectra.inter_spike_spectrum — Method

hallo 2

InterSpikeSpectra.inter_spike_spectrum2 — Method

hallo

InterSpikeSpectra.inter_spike_spectrum3 — Method

inter_spike_spectrum(s::Vector; kwargs...) → spectrum, ρ

Compute a spike `spectrum` from the a signal `s`, using a LASSO optimization.
`s` will get normalized to the unity interval `s ∈ [0 1]` internally. Second
output `ρ` is the linear Pearson correlation coefficient between the true
input `s` and the regenerated decomposed signal.

Keyword arguments:
`ρ_thres = 0.99`: The agreement of the regenerated decomposed signal with the
                  true signal. This depends on the LASSO regularization parameter
                  `λ`. `λ` gets adjusted automatically with respect to `ρ_thres`.
`tol = 1e-3`: Allowed tolerance between `ρ_thres` and `ρ`.
`max_iter = 15`: Determines after how many tried Lambdas the algorithm stopps.
`verbose::Bool=true`: If true, warning messages enabled.